refactor(agents): split transcript image redaction helpers
